+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package are documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2025-08-28
+### Added
+ - Added ```DbgDraw.SphereCast```.
+ - Added ```DbgDraw.BoxCast```.
+ - Added ```DbgDraw.WireSphereCast```.
+ - Added ```DbgDraw.WireBoxCast```.
+ - Added ```DbgDraw.WireCapsuleCast```.
+ 
 ## [1.2.0] - 2024-08-12
 ### Fixed
  - Fixed ```ArgumentNullException``` during internal ```DbgDraw``` initialization due to missing a required shader. I've implemented a build pre-processor that automatically adds the required ```Hidden/DbgDraw-Shaded``` shader to the 'Always Included Shaders' list found under 'Project Settings > Graphics'. This fixes issue #6.